% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_b1_correct.R
\name{mp2rage_b1_correct}
\alias{mp2rage_b1_correct}
\title{mp2rage_b1_correct}
\usage{
mp2rage_b1_correct(
  in_b1_map_data,
  is_sa2rage = TRUE,
  param_list_mp2rage,
  in_uni_data,
  param_list_sa2rage,
  inversion_efficiency = 0.96,
  b0 = 7
)
}
\arguments{
\item{in_b1_map_data}{(required) path to the B1 map NIfTI file (must be co-registered to UNI)}

\item{is_sa2rage}{(default) boolean \code{TRUE} by default, must be \code{FALSE} if TFL B1 map}

\item{param_list_mp2rage}{(required) list of parameters of MP2RAGE acquisition}

\item{in_uni_data}{(required) path to the magnitude MP2RAGE UNI NIfTI file}

\item{param_list_sa2rage}{(required if \code{is_sa2rage=TRUE}) list of parameters of Sa2RAGE acquisition}

\item{inversion_efficiency}{(default) 0.96 is the efficiency of the Siemens MP2RAGE inversion pulse}

\item{b0}{(has default) \emph{7.0} can be \emph{3.0}}
}
\value{
List of corrected B1, T1 and UNI numerical arrays
}
\description{
\code{mp2rage_b1_correct} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{MATLAB code} made publicly available by José P. Marques.

For methodological details, see \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE}, \href{https://onlinelibrary.wiley.com/doi/full/10.1002/mrm.23145}{Sa2RAGE} and \href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0069294}{Sa2RAGE correction}.

Please cite the linked papers if you used these methods in your work.
}
\examples{

mp2rage_params <- list(
                   mprage_tr = 5.0,
                   flash_tr = 6.9e-3,
                   inv_times_a_b = c(900e-3,2750e-3),
                   flip_angle_a_b_deg = c(5,3),
                   num_z_slices = c(120,120)
                   )

sa2rage_params <- list(
                   mprage_tr = 2.4,
                   flash_tr = 2.2e-3,
                   inv_times_a_b = c(58e-3,1800e-3),
                   flip_angle_a_b_deg = c(4,10),
                   num_z_slices = c(22,38)
                   )

\dontrun{
out_list <- mp2rage_b1_correct(
                   in_b1_map_data = "sub-01_Sa2RAGE_B1map_reg-to-MP2RAGE.nii.gz",
                   is_sa2rage = TRUE,
                   param_list_sa2rage = sa2rage_params,
                   in_uni_data = "sub-01_MP2RAGE_UNI.nii.gz",
                   param_list_mp2rage = mp2rage_params,
                   inversion_efficiency = 0.96,
                   b0 = 7
                   )

b1_corr_data <- out_list$b1_corr
t1_corr_data <- out_list$t1_corr
uni_corr_data <- out_list$uni_corr}
}
\author{
Sriranga Kashyap
}
