% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_b1_correction_short.R
\name{mp2rage_b1_correction_short}
\alias{mp2rage_b1_correction_short}
\title{mp2rage_b1_correction_short}
\usage{
mp2rage_b1_correction_short(
  in_b1_map,
  is_sa2rage = TRUE,
  in_uni,
  param_list_sa2rage,
  param_list_mp2rage,
  inversion_efficiency = 0.96
)
}
\value{
Data as a numerical array
}
\description{
\code{mp2rage_b1_correction_short} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of MATLAB code made publicly available by \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{José P. Marques}.

For methodological details, see \href{https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.1910150117}{MPRAGE paper} and \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE paper}.
}
\examples{
list_of_b1_corr_t1_corr_uni_corr <- mp2rage_b1_correction_short(in_b1_map,
is_sa2rage = TRUE,
in_uni,
param_list_sa2rage,
param_list_mp2rage,
inversion_efficiency = 0.96)
}
\author{
Sriranga Kashyap
}
