% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_estimate_t1_m0.R
\name{mp2rage_estimate_t1_m0}
\alias{mp2rage_estimate_t1_m0}
\title{mp2rage_estimate_t1_m0}
\usage{
mp2rage_estimate_t1_m0(in_uni_data, in_inv2_data = NULL, param_list_mp2rage)
}
\arguments{
\item{in_uni_data}{(required) path to the magnitude MP2RAGE UNI NIfTI file}

\item{in_inv2_data}{(has default) path to the magnitude MP2RAGE INV2 NIfTI file, not required for only T1 estimation}

\item{param_list_mp2rage}{(required) list of parameters of MP2RAGE acquisition}
}
\value{
List of numerical arrays (not NIfTI)
}
\description{
\code{mp2rage_estimate_t1_m0} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{MATLAB code} made publicly available by José P. Marques.

For methodological details, see \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE}.

Please cite the linked papers if you used these methods in your work.
}
\examples{
\dontrun{
list_t1_m0_maps <- mp2rage_estimate_t1_m0(
                         in_uni_data,
                         in_inv2_data,
                         param_list_mp2rage)}
}
\author{
Sriranga Kashyap
}
