% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage.R
\name{mp2rage_func}
\alias{mp2rage_func}
\title{mprage_func}
\usage{
mp2rage_func(
  mprage_tr = NULL,
  inv_times_a_b = NULL,
  flip_angle_a_b_deg = NULL,
  num_z_slices = NULL,
  flash_tr = NULL,
  sequence_type = "normal",
  inversion_efficiency = 0.96,
  b0 = 7,
  inv1_mag = NULL,
  inv1_phs = NULL,
  inv2_mag = NULL,
  inv2_phs = NULL,
  b1_map = NULL
)
}
\value{
Data as a numerical array
}
\description{
\code{mprage_func} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of MATLAB code made publicly available by \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{José P. Marques}.

For methodological details, see \href{https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.1910150117}{MPRAGE paper} and \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE paper}.
}
\examples{
Signal <- mprage_func(mprage_tr,
                      inv_times_a_b,
                      num_z_slices,
                      flash_tr,
                      flip_angle_a_b_deg,
                      sequence_type="normal",
                      t1s,
                      ...)
}
\author{
Sriranga Kashyap
}
