% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_interp_1d.R
\name{mp2rage_interp_1d}
\alias{mp2rage_interp_1d}
\title{mp2rage_interp_1d}
\usage{
mp2rage_interp_1d(a, b, a_int, is_linear = TRUE)
}
\value{
Data as a numerical array
}
\description{
\code{mp2rage_interp_1d} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of MATLAB code made publicly available by \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{José P. Marques}.

For methodological details, see \href{https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.1910150117}{MPRAGE paper} and \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE paper}.
}
\examples{
bi <- mp2rage_interp_1d(a, b, a_int, is_linear = TRUE)
}
\author{
Sriranga Kashyap
}
