% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_lookuptable.R
\name{mp2rage_lookuptable}
\alias{mp2rage_lookuptable}
\title{mp2rage_lookuptable}
\usage{
mp2rage_lookuptable(
  mp2rage_tr,
  inv_times_a_b,
  flip_angle_a_b_deg,
  num_z_slices,
  flash_tr,
  sequence_type,
  n_images = 2,
  b0 = 7,
  m0 = 1,
  inversion_efficiency = 0.96,
  all_data = 0,
  t1_vector = NULL
)
}
\value{
List of numerical arrays (in order): intensity, t1_vector, intensity_before_comb
}
\description{
\code{mp2rage_lookuptable} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of MaTLab code made publicly
available by \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{José P. Marques}.

For methodological details, see \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE paper}.

Please cite the aforementioned paper if you used this method in your work.
}
\examples{
list_of_intensity_t1_vector_intensity_before_comb <- mp2rage_lookuptable(mp2rage_tr,
                                                                         inv_times_a_b,
                                                                         flip_angle_a_b_deg,
                                                                         num_z_slices,
                                                                         flash_tr,
                                                                         sequence_type,
                                                                         ...)
}
\author{
Sriranga Kashyap
}
