% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_register_sa2rage.R
\name{mp2rage_register_sa2rage}
\alias{mp2rage_register_sa2rage}
\title{mp2rage_register_sa2rage}
\usage{
mp2rage_register_sa2rage(
  in_sa2rage_inv2,
  in_sa2rage_b1_map,
  in_mp2rage_inv2,
  out_reg_b1_map
)
}
\arguments{
\item{in_sa2rage_inv2}{(required) path to the magnitude Sa2RAGE INV2 NIfTI file}

\item{in_sa2rage_b1_map}{(required) path to the magnitude Sa2RAGE B1 map NIfTI file}

\item{in_mp2rage_inv2}{(required) path to the magnitude MP2RAGE INV2 NIfTI file}

\item{out_reg_b1_map}{(required) path to the registered Sa2RAGE B1 map NIfTI file}
}
\value{
Saves Sa2RAGE B1 map NIfTI file registered and resampled into MP2RAGE space.
}
\description{
\code{mp2rage_register_sa2rage} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{MATLAB code} made publicly available by José P. Marques.

Uses \href{https://github.com/jonclayden/RNifti}{RNifti} and \href{https://github.com/jonclayden/RNiftyReg}{RNiftyReg}.

Please cite the linked papers if you used these methods in your work.
}
\examples{
\dontrun{
mp2rage_register_sa2rage(
       in_sa2rage_inv2 = "sub-01_Sa2RAGE_INV2.nii.gz",
       in_sa2rage_b1_map = "sub-01_Sa2RAGE_B1map.nii.gz",
       in_mp2rage_inv2 = "sub-01_MP2RAGE_INV2.nii.gz",
       out_reg_b1_map = "sub-01_Sa2RAGE_B1map_reg-to-mp2rage.nii.gz")}
}
\author{
Sriranga Kashyap
}
