% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mprage_func.R
\name{mprage_func}
\alias{mprage_func}
\title{mprage_func}
\usage{
mprage_func(
  mprage_tr,
  inv_times_a_b,
  num_z_slices,
  flash_tr,
  flip_angle_a_b_deg,
  sequence_type = "normal",
  t1s,
  n_images = 2,
  b0 = 7,
  m0 = 1,
  inversion_efficiency = 0.96
)
}
\value{
Data as a numerical array
}
\description{
\code{mprage_func} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of MATLAB code made publicly available by \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{José P. Marques}.

For methodological details, see \href{https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.1910150117}{MPRAGE paper} and \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE paper}.
}
\examples{
Signal <- mprage_func(mprage_tr,
                      inv_times_a_b,
                      num_z_slices,
                      flash_tr,
                      flip_angle_a_b_deg,
                      sequence_type="normal",
                      t1s,
                      ...)
}
\author{
Sriranga Kashyap
}
