% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_combination.R
\name{robust_combination}
\alias{robust_combination}
\title{robust_combination}
\usage{
robust_combination(inv1, inv2, uni, uni_out = NULL, regularisation = 1)
}
\arguments{
\item{inv1}{path to MP2RAGE inv1 NIfTI file (required)}

\item{inv2}{path to MP2RAGE inv1 NIfTI file (required)}

\item{uni}{path to MP2RAGE uni NIfTI file (required)}

\item{uni_out}{(optional)  path to background suppressed MP2RAGE uni NIfTI file}

\item{regularisation}{(optional) a scalar (1-10) for background noise supression strength}
}
\value{
Data as a numerical array (not NIfTI, unless uni_out is specified)
}
\description{
\code{robust_combination} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of MATLAB code made publicly available by \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{José P. Marques}.

For methodological details, see \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE sequence} and \href{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0099676}{MP2RAGE background supression algorithm}.

Please cite the aforementioned papers if you used these methods in your work.
}
\examples{
mp2rage_robust_phase_sensitive <- robustCombination("sub-01_inv1.nii.gz","sub-01_inv2.nii.gz","sub-01_uni.nii.gz","sub-01_uni_out.nii.gz",regularisation = 10)
}
\author{
Sriranga Kashyap
}
