% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp2rage_bloch_func.R
\name{mp2rage_bloch_func}
\alias{mp2rage_bloch_func}
\title{mp2rage_bloch_func}
\usage{
mp2rage_bloch_func(
  mprage_tr,
  flash_tr,
  inv_times_a_b,
  flip_angle_a_b_deg,
  num_z_slices,
  sequence_type = "normal",
  t1_vector = NULL,
  b0 = 7,
  m0 = 1,
  inversion_efficiency = 0.96,
  n_images = 2
)
}
\arguments{
\item{mprage_tr}{(required) \emph{TR} of the MP2RAGE in s}

\item{flash_tr}{(required) \emph{TR} of the FLASH readout in s}

\item{inv_times_a_b}{(required) \eqn{c(TI1,TI2)}}

\item{flip_angle_a_b_deg}{(required) \emph{c(\eqn{\alpha}1,\eqn{\alpha}2)}}

\item{num_z_slices}{(required) calculate as: \eqn{slices_per_slab * c(slice_partial_fourier-0.5,0.5)}}

\item{sequence_type}{(has default) \emph{"normal"}, set to \emph{NULL} for water excitation}

\item{t1_vector}{(has default) \emph{NULL}, provide input vector as necessary}

\item{b0}{(has default) \emph{7.0} can be \emph{3.0}}

\item{m0}{(has default) \emph{1.0}}

\item{inversion_efficiency}{(has default) \emph{0.96} is the efficiency of the Siemens MP2RAGE inversion pulse}

\item{n_images}{(has default) \emph{2}, change if required}
}
\value{
Data as a numerical array
}
\description{
\code{mp2rage_bloch_func} is part of the \pkg{mp2rageR} package.

\pkg{mp2rageR} is the R implementation of \href{https://github.com/JosePMarques/MP2RAGE-related-scripts}{MATLAB code} made publicly available by José P. Marques.

For methodological details, see \href{https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.1910150117}{MPRAGE} and \href{https://www.sciencedirect.com/science/article/abs/pii/S1053811909010738}{MP2RAGE}.

Please cite the linked papers if you used these methods in your work.
}
\examples{
mp2rage_params <- list(
                   mprage_tr = 5.0,
                   flash_tr = 6.9e-3,
                   inv_times_a_b = c(900e-3,2750e-3),
                   flip_angle_a_b_deg = c(5,3),
                   num_z_slices = c(120,120)
                   )

\dontrun{
signal <- mp2rage_bloch_func(
                   mp2rage_params$mprage_tr,
                   mp2rage_params$flash_tr,
                   mp2rage_params$inv_times_a_b,
                   mp2rage_params$flip_angle_a_b_deg,
                   mp2rage_params$num_z_slices,
                   sequence_type = "normal",
                   t1_vector = NULL,
                   b0 = 7,
                   m0 = 1,
                   inversion_efficiency = 0.96,
                   n_images = 2
                   )}
}
\author{
Sriranga Kashyap
}
